+++ core/api_container/server/service_network/mock_service_network.go
@@ -9,7 +9,6 @@ import (
 	"bytes"
 	"context"
 	"github.com/docker/go-connections/nat"
-	"github.com/kurtosis-tech/kurtosis-client/golang/kurtosis_core_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis/api_container/server/service_network/partition_topology"
 	"github.com/kurtosis-tech/kurtosis/api_container/server/service_network/service_network_types"
 	"github.com/palantir/stacktrace"
@@ -30,15 +29,7 @@ func (m MockServiceNetwork) Repartition(ctx context.Context, newPartitionService
 	panic("This is unimplemented for the mock network")
 }
 
-func (m MockServiceNetwork) RegisterService(serviceId service_network_types.ServiceID, partitionId service_network_types.PartitionID) (net.IP, error) {
-	panic("This is unimplemented for the mock network")
-}
-
-func (m MockServiceNetwork) GenerateFiles(serviceId service_network_types.ServiceID, filesToGenerate map[string]*kurtosis_core_rpc_api_bindings.FileGenerationOptions) (map[string]string, error) {
-	panic("This is unimplemented for the mock network")
-}
-
-func (m MockServiceNetwork) LoadStaticFiles(serviceId service_network_types.ServiceID, staticFileIdKeys map[string]bool) (map[string]string, error) {
+func (m MockServiceNetwork) RegisterService(serviceId service_network_types.ServiceID, partitionId service_network_types.PartitionID) (net.IP, string, error) {
 	panic("This is unimplemented for the mock network")
 }
 
@@ -68,6 +59,10 @@ func (m MockServiceNetwork) GetServiceEnclaveDataVolMntDirpath(serviceId service
 		return "", stacktrace.NewError("No volume directory path defined for service with ID '%v'", serviceId)
 	}
 	return volMntDirPath, nil
+}
+
+func (m MockServiceNetwork) GetRelativeServiceDirpath(serviceId service_network_types.ServiceID) (string, error) {
+	panic("This is unimplemented for the mock network")
 }
 
 func (m MockServiceNetwork) Destroy(ctx context.Context) error {